+++ spec/client/group_client_spec.rb
@@ -50,4 +50,38 @@ RSpec.describe KeycloakAdmin::GroupClient do
       expect(groups[0].name).to eq "test_group_name"
     end
   end
+
+  describe "#save" do
+    let(:realm_name) { "valid-realm" }
+    let(:group) { KeycloakAdmin::GroupRepresentation.from_hash(
+      "name" => "test_group_name"
+    )}
+
+    before(:each) do
+      @group_client = KeycloakAdmin.realm(realm_name).groups
+
+      stub_token_client
+      response = double
+      allow(response).to receive(:headers).and_return(
+        { location: 'http://auth.service.io/auth/admin/realms/valid-realm/groups/be061c48-6edd-4783-a726-1a57d4bfa22b' }
+      )
+      allow_any_instance_of(RestClient::Resource).to receive(:post).and_return response
+    end
+
+    it "saves a group" do
+      group_id = @group_client.save(group)
+      expect(group_id).to eq 'be061c48-6edd-4783-a726-1a57d4bfa22b'
+    end
+
+    it "passes rest client options" do
+      rest_client_options = {verify_ssl: OpenSSL::SSL::VERIFY_NONE}
+      allow_any_instance_of(KeycloakAdmin::Configuration).to receive(:rest_client_options).and_return rest_client_options
+
+      expect(RestClient::Resource).to receive(:new).with(
+        "http://auth.service.io/auth/admin/realms/valid-realm/groups", rest_client_options).and_call_original
+
+      group_id = @group_client.save(group)
+      expect(group_id).to eq 'be061c48-6edd-4783-a726-1a57d4bfa22b'
+    end
+  end
 end